Hang

+++ code/DynamicStereo/dynamicstereo.h
@@ -22,7 +22,7 @@
 namespace dynamic_stereo {
     class DynamicStereo {
     public:
-	    typedef int EnergyType;
+	    typedef double EnergyType;
 
         DynamicStereo(const FileIO& file_io_, const int anchor_, const int tWindow_, const int tWindowStereo_, const int downsample_, const double weight_smooth_,
                       const int dispResolution_ = 64, const double min_disp_ = -1, const double max_disp_ = -1);
